requirements.txt
Flask
Flask-WTF
python-dotenv
openai>=1.14
gunicorn
